# Optimize get_inputs_embeds() for Qwen2VL.