--- a/src/cpp/src/icontinuous_batching.cpp
+++ b/src/cpp/src/icontinuous_batching.cpp
@@ -191,7 +191,6 @@ ContinuousBatchingPipeline::IContinuousBatchingPipeline::generate(
 
         ov::Tensor embeds;
         if (qwen2vl_vision_encoder && m_history_images.size() > 0) {
-            // If m_merged_image_embeddings is not passed it will be recalculated each time
             embeds = m_inputs_embedder->get_inputs_embeds(templated_history, m_history_images, vlm_perf_metrics[0], m_merged_image_embeddings);
         } else {
             embeds = m_inputs_embedder->get_inputs_embeds(templated_history, m_history_images, vlm_perf_metrics[0]);