… src/cpp/src/visual_language/minicpm/classes.cpp
… src/cpp/src/visual_language/minicpm/classes.hpp
… src/cpp/src/visual_language/phi3_vision/classes.cpp
… src/cpp/src/visual_language/phi3_vision/classes.hpp
… src/cpp/src/visual_language/qwen2vl/classes.cpp
… src/cpp/src/visual_language/qwen2vl/classes.hpp