Image embeddings merger is moved to a separate method and in chat mode it is used only when new images are passed.
Time measures are below for 6 chat iterations for Qwen2-VL-2B-Instruct, where image is passed on the first and third iteration.

This branch:
Chat iteration 1 (new image):
encode time: 2012 ms
get_inputs_embeds time: 7683 ms
Chat iteration 2:
encode time: 0 ms
get_inputs_embeds time: 7 ms
Chat iteration 3 (new image):
encode time: 2359 ms
get_inputs_embeds time: 29179 ms
Chat iteration 4:
encode time: 0 ms
get_inputs_embeds time: 10 ms
Chat iteration 5:
encode time: 0 ms
get_inputs_embeds time: 11 ms
Chat iteration 6:
encode time: 0 ms
get_inputs_embeds time: 8 ms



Master:
Chat iteration 1 (new image):
encode time: 1893ms
get_inputs_embeds time: 8394ms
Chat iteration 2:
encode time: 0ms
get_inputs_embeds time: 7664ms
Chat iteration 3 (new image):
encode time: 2126ms
get_inputs_embeds time: 27954ms
Chat iteration 4:
encode time: 0ms
get_inputs_embeds time: 27944ms
Chat iteration 5:
encode time: 0ms
get_inputs_embeds time: 27974ms
Chat iteration 6:
encode time: 0ms
get_inputs_embeds time: 27970ms