@Wovchena 2025-04-11
Seems like true positive in `Python (LLM & VLM) Tests (wheel)`:
`/Users/runner/work/_temp/832cc46c-22c3-4d08-b3ef-d88ed61bdf98.sh: line 1:  5146 Segmentation fault: 11  python -m pytest -v ./tests/python_tests/test_llm_pipeline.py tests/python_tests/test_llm_pipeline_static.py tests/python_tests/test_vlm_pipeline.py
tests/python_tests/test_vlm_pipeline.py::TestImageTags::test_same_reference[model_to_tag0] `
0 corresponds to `katuni4ka/tiny-random-qwen2vl`