@Wovchena 2025-04-21
`tests/python_tests/test_vlm_pipeline.py::TestImageTags::test_append[model_to_tag1]` fails in other PRs as well